% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob.R
\name{calc_prob}
\alias{calc_prob}
\title{Receive a quantile of a normal distribuition(0;1),
calculate this density and converts this valor in percentual}
\usage{
calc_prob(x)
}
\arguments{
\item{x}{quantile of normal}
}
\value{
percentual number
}
\description{
Receive a quantile of a normal distribuition(0;1),
calculate this density and converts this valor in percentual
}
\examples{
calc_prob(0.7)
}
