% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{converter}
\alias{converter}
\title{Converter}
\usage{
converter(a)
}
\arguments{
\item{a}{number type double}
}
\value{
percentage of argument 'a'
}
\description{
Convert a probability for a percent number
and return it
}
\examples{
converter(dnorm(0.3))
}
